% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{RcaCovidVaccine-package}
\alias{RcaCovidVaccine}
\alias{RcaCovidVaccine-package}
\title{RcaCovidVaccine: Safety Analysis for COVID-19 Vaccines}
\description{
Safety analysis for COVID-19 vaccines.
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Patrick Ryan
}

}
\keyword{internal}
