% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the Study}
\usage{
execute(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  verifyDependencies = TRUE,
  outputFolder,
  databaseId = "Unknown",
  databaseName = "Unknown",
  databaseDescription = "Unknown",
  createCohorts = TRUE,
  runCohortDiagnostics = TRUE,
  runCohortMethod = TRUE,
  runSccs = TRUE,
  generateCohortMethodDiagnostics = TRUE,
  generatSccsDiagnostics = TRUE,
  maxCores = 4,
  minCellCount = 5
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write privilege in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{tempEmulationSchema}{Some database platforms like Oracle and Impala do not truly support temp tables. To
emulate temp tables, provide a schema with write privileges where temp tables
can be created.}

\item{verifyDependencies}{Check whether correct package versions are installed?}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{databaseId}{A short string for identifying the database (e.g.
'Synpuf').}

\item{databaseName}{The full name of the database (e.g. 'Medicare Claims
Synthetic Public Use Files (SynPUFs)').}

\item{databaseDescription}{A short description (several sentences) of the database.}

\item{createCohorts}{Create the cohortTable table with the exposure and outcome cohorts?}

\item{runCohortDiagnostics}{Perform cohort diagnostics?}

\item{runCohortMethod}{Perform the cohort method analyses?}

\item{runSccs}{Perform the SCCS analyses?}

\item{generateCohortMethodDiagnostics}{Generate diagnostics for the cohort method analyses?}

\item{generatSccsDiagnostics}{Generate diagnostics for the SCCS analyses?}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}

\item{minCellCount}{The minimum number of subjects contributing to a count before it can be included
in packaged results.}
}
\description{
Execute the Study
}
\details{
This function executes the RcaCovidVaccine Study.
}
