% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unblind.R
\name{unblind}
\alias{unblind}
\title{Unblind results}
\usage{
unblind(
  verifyDependencies = TRUE,
  outputFolder,
  databaseId,
  unblindCohortMethod = TRUE,
  unblindSccs = TRUE
)
}
\arguments{
\item{verifyDependencies}{Check whether correct package versions are installed?}

\item{outputFolder}{Name of local folder containing the results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{databaseId}{A short string for identifying the database (e.g.
'Synpuf').}

\item{unblindCohortMethod}{Unblind the cohort method analyses?}

\item{unblindSccs}{Unblind the SCCS analyses?}
}
\description{
Unblind results
}
\details{
Unblind the results of analyses that have passed diagnostics.
}
